Fix for common name display for previously seen organisms on Match
