+++ package.json
@@ -67,8 +67,7 @@
     "@sayem314/react-native-keep-awake": "^1.2.3",
     "@shopify/flash-list": "^1.7.2",
     "@tanstack/react-query": "^5.29.0",
-    "apisauce": "^3.0.1",
-    "axios": "^1.7.4",
+    "apisauce": "^3.1.0",
     "classnames": "^2.5.1",
     "date-fns": "^3.0.0",
     "date-fns-tz": "^3.0.0",
@@ -168,7 +167,7 @@
     "babel-plugin-transform-remove-console": "^6.9.4",
     "chalk": "^5.4.1",
     "decompress": "^4.2.1",
-    "detox": "^20.32.0",
+    "detox": "^20.34.5",
     "dotenv": "^16.4.7",
     "eslint": "^8.57.0",
     "eslint-config-airbnb": "^19.0.4",
